02-minicurso_junit/src/main/java/dcomp/es2/locadora/modelo/Filme.java
package dcomp.es2.locadora.modelo;

public class Filme {

	private String nome;
	private Integer estoque;
	private Double precoLocacao;  
	
	public Filme() {}
	
	public Filme(String nome, Integer estoque, Double precoLocacao) {
		this.nome = nome;
		this.estoque = estoque;
		this.precoLocacao = precoLocacao;
	}
	
	public String getNome() {
		return nome;
	}
	public void setNome(String nome) {
		this.nome = nome;
	}
	public Integer getEstoque() {
		return estoque;
	}
	public void setEstoque(Integer estoque) {
		this.estoque = estoque;
	}
	public Double getPrecoLocacao() {
		return precoLocacao;
	}

	public void setPrecoLocacao(Double precoLocacao) {
		if (precoLocacao <= 0 ) {
			throw new IllegalArgumentException("O preço da locaçao deve ser maior do que Zero");
		}

		this.precoLocacao = precoLocacao;
	}
}